+++ api/handlers.py
@@ -1,5 +1,4 @@
 from logging import getLogger
-from typing import Union
 from uuid import UUID
 
 from fastapi import APIRouter
@@ -8,122 +7,84 @@ from fastapi import HTTPException
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from api.actions.auth import get_current_user_from_token
+from api.actions.user import _create_new_user
+from api.actions.user import _delete_user
+from api.actions.user import _get_user_by_id
+from api.actions.user import _update_user
 from api.models import DeleteUserResponse
 from api.models import ShowUser
 from api.models import UpdatedUserResponse
 from api.models import UpdateUserRequest
 from api.models import UserCreate
-from db.dals import UserDAL
+from db.models import User
 from db.session import get_db
-from hashing import Hasher
-
 
 logger = getLogger(__name__)
 
 user_router = APIRouter()
 
 
-@user_router.get("/get_user", response_model=ShowUser)
-async def get_user_by_id(user_id: UUID, db: AsyncSession = Depends(get_db)) -> Union:
-    user = await _get_user_by_id(user_id, db)
-    if user is None:
-        raise HTTPException(status_code=404, detail=f"User with id {user_id} not found")
-    return user
-
-
-@user_router.post("/create", response_model=ShowUser)
+@user_router.post("/", response_model=ShowUser)
 async def create_user(body: UserCreate, db: AsyncSession = Depends(get_db)) -> ShowUser:
     try:
         return await _create_new_user(body, db)
-    except IntegrityError as error:
-        raise HTTPException(status_code=503, detail=f"database error: {error}")
+    except IntegrityError as err:
+        logger.error(err)
+        raise HTTPException(status_code=503, detail=f"Database error: {err}")
 
 
-@user_router.delete("/delete", response_model=DeleteUserResponse)
+@user_router.delete("/", response_model=DeleteUserResponse)
 async def delete_user(
-    user_id: UUID, db: AsyncSession = Depends(get_db)
+    user_id: UUID,
+    db: AsyncSession = Depends(get_db),
+    current_user: User = Depends(get_current_user_from_token),
 ) -> DeleteUserResponse:
     deleted_user_id = await _delete_user(user_id, db)
     if deleted_user_id is None:
-        raise HTTPException(status_code=404, detail=f"user with {user_id} not found")
+        raise HTTPException(
+            status_code=404, detail=f"User with id {user_id} not found."
+        )
     return DeleteUserResponse(deleted_user_id=deleted_user_id)
 
 
-@user_router.patch("/patch", response_model=UpdatedUserResponse)
+@user_router.get("/", response_model=ShowUser)
+async def get_user_by_id(
+    user_id: UUID,
+    db: AsyncSession = Depends(get_db),
+    current_user: User = Depends(get_current_user_from_token),
+) -> ShowUser:
+    user = await _get_user_by_id(user_id, db)
+    if user is None:
+        raise HTTPException(
+            status_code=404, detail=f"User with id {user_id} not found."
+        )
+    return user
+
+
+@user_router.patch("/", response_model=UpdatedUserResponse)
 async def update_user_by_id(
-    user_id: UUID, body: UpdateUserRequest, db: AsyncSession = Depends(get_db)
+    user_id: UUID,
+    body: UpdateUserRequest,
+    db: AsyncSession = Depends(get_db),
+    current_user: User = Depends(get_current_user_from_token),
 ) -> UpdatedUserResponse:
-    updated_users_params = body.dict(exclude_none=True)
-    if updated_users_params == {}:
+    updated_user_params = body.dict(exclude_none=True)
+    if updated_user_params == {}:
         raise HTTPException(
             status_code=422,
-            detail="At least one parameter for user update info should provided",
+            detail="At least one parameter for user update info should be provided",
         )
     user = await _get_user_by_id(user_id, db)
     if user is None:
         raise HTTPException(
-            status_code=404, detail=f"User with user_id {user_id} not found"
+            status_code=404, detail=f"User with id {user_id} not found."
         )
     try:
         updated_user_id = await _update_user(
-            updated_user_params=updated_users_params, db=db, user_id=user_id
+            updated_user_params=updated_user_params, session=db, user_id=user_id
         )
-    except IntegrityError as error:
-        raise HTTPException(status_code=503, detail=f"database error:{error}")
+    except IntegrityError as err:
+        logger.error(err)
+        raise HTTPException(status_code=503, detail=f"Database error: {err}")
     return UpdatedUserResponse(updated_user_id=updated_user_id)
-
-
-async def _create_new_user(body: UserCreate, db) -> ShowUser:
-    async with db as session:
-        async with session.begin():
-            user_dal = UserDAL(session)
-            user = await user_dal.create_user(
-                name=body.name,
-                surname=body.surname,
-                email=body.email,
-                hashed_password=Hasher.get_password_hash(body.password),
-            )
-            return ShowUser(
-                user_id=user.user_id,
-                name=user.name,
-                surname=user.surname,
-                email=user.email,
-                is_active=user.is_active,
-            )
-
-
-async def _delete_user(user_id, db) -> Union[UUID, None]:
-    async with db as session:
-        async with session.begin():
-            user_dal = UserDAL(session)
-            deleted_user_id = await user_dal.delete_user(
-                user_id=user_id,
-            )
-            return deleted_user_id
-
-
-async def _get_user_by_id(user_id, db) -> Union[ShowUser, None]:
-    async with db as session:
-        async with session.begin():
-            user_dal = UserDAL(session)
-            selected_user = await user_dal.get_user_by_id(user_id=user_id)
-            if selected_user is not None:
-                return ShowUser(
-                    user_id=selected_user.user_id,
-                    name=selected_user.name,
-                    surname=selected_user.surname,
-                    email=selected_user.email,
-                    is_active=selected_user.is_active,
-                )
-
-
-async def _update_user(
-    updated_user_params: dict, user_id: UUID, db
-) -> Union[UUID, None]:
-    async with db as session:
-        async with session.begin():
-            user_dal = UserDAL(session)
-            updated_user_id = await user_dal.update_user(
-                user_id=user_id, **updated_user_params
-            )
-        return updated_user_id
